# Conflicts:
#	src/main/java/com/qiein/jupiter/web/controller/ReportsController.java
#	src/main/java/com/qiein/jupiter/web/service/ReportService.java
#	src/main/java/com/qiein/jupiter/web/service/impl/ReportsServiceImpl.java

+++ src/main/java/com/qiein/jupiter/web/repository/SourceOrderDataReportsDao.java
@@ -280,7 +280,7 @@ public class SourceOrderDataReportsDao {
         allReportsVO.setSrcType(-1);
         allReportsVO.setSrcName(CommonConstant.totalAll);
         //总销售均价
-        allReportsVO.setSaleAvg(NumUtil.keep2Point(allReportsVO.getTxTotalAmount() / allReportsVO.getOrderAmount()));
+        allReportsVO.setSaleAvg(NumUtil.keep2Point((double)allReportsVO.getTxTotalAmount() / allReportsVO.getOrderAmount()));
 
         reportsList.add(0, allReportsVO);
     }
